unit tests from cosmetic.rs moved to tests/unit/filters/cosmetic.rs
